# generated/openapi-admin.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/rental-orders": {
      "get": {
        "operationId": "RentalController_list[1]",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "RESERVED",
                "CONFIRMED",
                "ACTIVE",
                "RETURNED",
                "COMPLETED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "paymentStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "UNPAID",
                "PARTIALLY_PAID",
                "PAID",
                "REFUNDED",
                "PARTIALLY_REFUNDED"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalOrderPageResDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List/search orders; use from/until for calendar overlap queries",
        "tags": [
          "Admin - Rental Orders"
        ]
      },
      "post": {
        "operationId": "RentalController_create[1]",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Recommended for retries from admin/FE.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRentalOrderReqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalOrderResDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create order + price snapshots + physical allocations atomically",
        "tags": [
          "Admin - Rental Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kitty Shop API - Admin API",
    "description": "Rental shop admin API for staff & operations (kitty-admin-fe). Tenant scope comes from the authenticated shop membership. Monetary values are decimal-safe values.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "API prefix"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Configured application URL"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RentalOrderPageResDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalOrderListItemResDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaResDto"
          }
        },
        "required": [
          "items",
          "meta"
        ]
      },
      "CreateRentalOrderReqDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string"
          },
          "locationId": {
            "type": "string"
          },
          "rentalStartAt": {
            "type": "string",
            "example": "2026-09-12T03:00:00.000Z"
          },
          "rentalEndAt": {
            "type": "string",
            "example": "2026-09-14T03:00:00.000Z"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateRentalItemReqDto"
            }
          },
          "charges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalChargeReqDto"
            }
          },
          "discountTotal": {
            "type": "number",
            "default": 0
          },
          "note": {
            "type": "string"
          },
          "internalNote": {
            "type": "string"
          },
          "delivery": {
            "$ref": "#/components/schemas/RentalDeliveryReqDto"
          },
          "collateral": {
            "$ref": "#/components/schemas/RentalCollateralReqDto"
          }
        },
        "required": [
          "customerId",
          "rentalStartAt",
          "rentalEndAt",
          "items"
        ]
      },
      "RentalOrderResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderNumber": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "rentalStartAt": {
            "type": "string",
            "format": "date-time"
          },
          "rentalEndAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "paymentStatus": {
            "type": "string"
          },
          "depositStatus": {
            "type": "string"
          },
          "grandTotal": {
            "type": "string",
            "example": "500000.00"
          },
          "customer": {
            "$ref": "#/components/schemas/RentalCustomerResDto"
          },
          "itemCount": {
            "type": "number"
          },
          "productCount": {
            "type": "number"
          },
          "confirmation": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/RentalConfirmationResDto"
              }
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalItemResDto"
            }
          },
          "rentalSubtotal": {
            "type": "string",
            "example": "450000.00"
          },
          "chargesTotal": {
            "type": "string",
            "example": "50000.00"
          },
          "discountTotal": {
            "type": "string",
            "example": "0.00"
          },
          "depositRequired": {
            "type": "string",
            "example": "1000000.00"
          },
          "collateralMethod": {
            "type": "string"
          },
          "documentType": {
            "type": "string",
            "nullable": true
          },
          "collateralStatus": {
            "type": "string"
          },
          "collateralReceivedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "collateralReturnedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "actualReturnedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "paidAmount": {
            "type": "string",
            "example": "250000.00"
          },
          "remainingAmount": {
            "type": "string",
            "example": "250000.00"
          },
          "settlement": {
            "$ref": "#/components/schemas/RentalSettlementResDto"
          },
          "returnRecord": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/RentalReturnResDto"
              }
            ]
          },
          "settlementDetails": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/RentalSettlementDetailsResDto"
              }
            ]
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "internalNote": {
            "type": "string",
            "nullable": true
          },
          "charges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalChargeResDto"
            }
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalPaymentResDto"
            }
          },
          "deliveries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalDeliveryResDto"
            }
          },
          "statusHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalStatusHistoryResDto"
            }
          }
        },
        "required": [
          "id",
          "orderNumber",
          "customerId",
          "rentalStartAt",
          "rentalEndAt",
          "status",
          "paymentStatus",
          "depositStatus",
          "grandTotal",
          "customer",
          "itemCount",
          "productCount",
          "confirmation",
          "items",
          "rentalSubtotal",
          "chargesTotal",
          "discountTotal",
          "depositRequired",
          "collateralMethod",
          "documentType",
          "collateralStatus",
          "collateralReceivedAt",
          "collateralReturnedAt",
          "actualReturnedAt",
          "paidAmount",
          "remainingAmount",
          "settlement",
          "returnRecord",
          "settlementDetails",
          "note",
          "internalNote",
          "charges",
          "payments",
          "deliveries",
          "statusHistory"
        ]
      },
...
      "PaginationMetaResDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
...
      "RentalOrderListItemResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderNumber": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "rentalStartAt": {
            "type": "string",
            "format": "date-time"
          },
          "rentalEndAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "paymentStatus": {
            "type": "string"
          },
          "depositStatus": {
            "type": "string"
          },
          "grandTotal": {
            "type": "string",
            "example": "500000.00"
          },
          "customer": {
            "$ref": "#/components/schemas/RentalCustomerResDto"
          },
          "itemCount": {
            "type": "number"
          },
          "productCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "orderNumber",
          "customerId",
          "rentalStartAt",
          "rentalEndAt",
          "status",
          "paymentStatus",
          "depositStatus",
          "grandTotal",
          "customer",
          "itemCount",
          "productCount"
        ]
      },
      "CreateRentalItemReqDto": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "default": 1,
            "minimum": 1,
            "maximum": 20
          },
          "unitRentalPrice": {
            "type": "number",
            "description": "Đơn giá thuê mỗi món do admin ghi đè (VNĐ). Nếu để trống, hệ thống dùng giá cấu hình.",
            "example": 150000
          },
          "inventoryItemIds": {
            "description": "Optional physical item selection; otherwise backend auto-allocates.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "variantId",
          "quantity"
        ]
      },
      "RentalChargeReqDto": {
        "type": "object",
        "properties": {
          "chargeType": {
            "type": "string",
            "enum": [
              "RENTAL_EXTRA",
              "ACCESSORY",
              "SHIPPING",
              "LATE",
              "CLEANING",
              "DAMAGE",
              "REPAIR",
              "LOST_ITEM",
              "OTHER"
            ],
            "example": "ACCESSORY"
          },
          "description": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "example": 50000
          },
          "quantity": {
            "type": "number",
            "default": 1
          }
        },
        "required": [
          "chargeType",
          "amount"
        ]
      },
      "RentalDeliveryReqDto": {
        "type": "object",
        "properties": {
          "direction": {
            "type": "string",
            "enum": [
              "OUTBOUND",
              "RETURN"
            ],
            "default": "OUTBOUND"
          },
          "method": {
            "type": "string",
            "enum": [
              "CUSTOMER_PICKUP",
              "SHOP_DELIVERY",
              "THIRD_PARTY_SHIPPER"
            ]
          },
          "scheduledAt": {
            "type": "string"
          },
          "recipientName": {
            "type": "string"
          },
          "recipientPhone": {
            "type": "string"
          },
          "addressLine": {
            "type": "string"
          },
          "ward": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "shippingFee": {
            "type": "number",
            "default": 0
          }
        },
        "required": [
          "direction",
          "method"
        ]
      },
      "RentalCollateralReqDto": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "CASH",
              "DOCUMENT"
            ]
          },
          "documentType": {
            "type": "string",
            "enum": [
              "CCCD",
              "GPLX"
            ]
          }
        },
        "required": [
          "method"
        ]
      },
      "RentalCustomerResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "fullName",
          "phone"
        ]
      },
      "RentalConfirmationResDto": {
        "type": "object",
        "properties": {
          "confirmedAt": {
            "type": "string",
            "format": "date-time"
          },
          "confirmedBy": {
            "type": "string"
          },
          "actorName": {
            "type": "string"
          },
          "rentalAmount": {
            "type": "string",
            "description": "Rental amount manually acknowledged; not a payment ledger entry."
          },
          "collateralMethod": {
            "type": "string"
          },
          "documentType": {
            "type": "string",
            "nullable": true
          },
          "collateralAmount": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "hasEvidence": {
            "type": "boolean"
          },
          "evidenceFilename": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "confirmedAt",
          "confirmedBy",
          "actorName",
          "rentalAmount",
          "collateralMethod",
          "documentType",
          "collateralAmount",
          "note",
          "hasEvidence",
          "evidenceFilename"
        ]
      },
      "RentalItemResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "productNameSnapshot": {
            "type": "string"
          },
          "variantNameSnapshot": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "productId": {
            "type": "string"
          },
          "variantId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "unitRentalPrice": {
            "type": "string"
          },
          "depositAmount": {
            "type": "string"
          },
          "lineTotal": {
            "type": "string"
          },
          "allocations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalAllocationResDto"
            }
          }
        },
        "required": [
          "id",
          "productNameSnapshot",
          "variantNameSnapshot",
          "quantity",
          "productId",
          "variantId",
          "status",
          "unitRentalPrice",
          "depositAmount",
          "lineTotal",
          "allocations"
        ]
      },
      "RentalSettlementResDto": {
        "type": "object",
        "properties": {
          "depositAvailable": {
            "type": "string"
          },
          "depositReceived": {
            "type": "string"
          },
          "refundAmount": {
            "type": "string"
          },
          "amountStillDue": {
            "type": "string"
          },
          "settlementStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "REFUND_DUE",
              "AMOUNT_DUE",
              "BALANCED",
              "SETTLED"
            ]
          }
        },
        "required": [
          "depositAvailable",
          "depositReceived",
          "refundAmount",
          "amountStillDue",
          "settlementStatus"
        ]
      },
      "RentalReturnResDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "returnedAt": {
            "type": "string",
            "format": "date-time"
          },
          "receivedBy": {
            "type": "string"
          },
          "actorName": {
            "type": "string",
            "nullable": true
          },
          "lateDays": {
            "type": "number"
          },
          "lateFee": {
            "type": "string",
            "example": "20000.00"
          },
          "additionalRentalFee": {
            "type": "string",
            "example": "0.00"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "inspections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentalReturnInspectionResDto"
            }
          }
        },
        "required": [
          "orderId",
          "returnedAt",
          "receivedBy",
          "actorName",
          "lateDays",
          "lateFee",
          "additionalRentalFee",
          "note",
          "inspections"
        ]
      },
      "RentalSettlementDetailsResDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "settledAt": {
            "type": "string",
            "format": "date-time"
          },
          "settledBy": {
            "type": "string"
          },
          "actorName": {
            "type": "string",
            "nullable": true
          },
          "settlementType": {
            "type": "string",
            "enum": [
              "REFUND",
              "COLLECTION",
              "BALANCED",
              "COLLATERAL_ONLY"
            ]
          },
          "amount": {
            "type": "string",
            "example": "0.00"
          },
          "depositAmount": {
            "type": "string",
            "example": "300000.00"
          },
          "totalCharges": {
            "type": "string",
            "example": "70000.00"
          },
          "refundAmount": {
            "type": "string",
            "example": "230000.00"
          },
          "amountDue": {
            "type": "string",
            "example": "0.00"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "evidenceKey": {
            "type": "string",
            "nullable": true
          },
          "evidenceFilename": {
            "type": "string",
            "nullable": true
          },
          "evidenceMimeType": {
            "type": "string",
            "nullable": true
          },
          "evidenceSize": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "orderId",
          "settledAt",
          "settledBy",
          "actorName",
          "settlementType",
          "amount",
          "depositAmount",
          "totalCharges",
          "refundAmount",
          "amountDue",
          "note",
          "evidenceKey",
          "evidenceFilename",
          "evidenceMimeType",
          "evidenceSize"
        ]
      },
      "RentalChargeResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "chargeType": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "chargeType",
          "description",
          "amount",
          "quantity",
          "currency"
        ]
      },
      "RentalPaymentResDto": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "transactionNumber": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "purpose": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "paidAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "source",
          "createdBy",
          "note",
          "id",
          "transactionNumber",
          "direction",
          "purpose",
          "paymentMethod",
          "amount",
          "currency",
          "paidAt"
        ]
      },
      "RentalDeliveryResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "recipientName": {
            "type": "string",
            "nullable": true
          },
          "recipientPhone": {
            "type": "string",
            "nullable": true
          },
          "addressLine": {
            "type": "string",
            "nullable": true
          },
          "shippingFee": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "direction",
          "method",
          "status",
          "scheduledAt",
          "recipientName",
          "recipientPhone",
          "addressLine",
          "shippingFee"
        ]
      },
      "RentalStatusHistoryResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fromStatus": {
            "type": "string",
            "nullable": true
          },
          "toStatus": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "changedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "fromStatus",
          "toStatus",
          "reason",
          "changedAt"
        ]
      },
...
      "RentalAllocationResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "inventoryItemId": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "operationalStatus": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "reservedFrom": {
            "type": "string",
            "format": "date-time"
          },
          "reservedUntil": {
            "type": "string",
            "format": "date-time"
          },
          "releasedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "inventoryItemId",
          "sku",
          "operationalStatus",
          "status",
          "reservedFrom",
          "reservedUntil",
          "releasedAt"
        ]
      },
      "RentalReturnInspectionResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "inventoryItemId": {
            "type": "string"
          },
          "condition": {
            "type": "string",
            "enum": [
              "NORMAL",
              "CLEANING_REQUIRED",
              "REPAIR_REQUIRED",
              "DAMAGED",
              "LOST"
            ]
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "inventoryItemId",
          "condition",
          "note"
        ]
      },
...
    }
  }
}